requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
anthropic==0.18.1
deepgram-sdk==3.2.7
supabase==2.3.4
pydantic[email]==2.5.0
PyMuPDF==1.23.8
python-docx==1.1.0
python-dotenv==1.0.0
